Do not filter completions based on prefix
Don't filter using 'toComplete' as a prefix because the shell can do
smarter filtering.  Zsh and fish in particular can do smart filtering
that is not solely based on prefix matching.  Top enable this shell
feature, we need to return all possile completions and let the shell
do the filtering itself.

A side-effect is that Cobra will not know in advance if some completions
will match or not, and won't be able to act based on that information.
It turns out there was only one case where this knowledge was needed
which was for ArgAliases.  However, the original behavior can be
achieved bby the program using ValidArgsFunction instead.

Signed-off-by: Marc Khouzam <marc.khouzam@montreal.ca>