Add QuickHelp for zsh and bash
QuickHelp allows the user to press a single key to obtain the help
text corresponding to the command-line, and then get back to that
command-line.

By default, the script maps F1 as the QuickHelp key.

For example:
   program completion<F1>
will show the help text of "program completion" by calling
"program completion --help".

The user can choose a different key combination for QuickHelp by setting
the QUICK_HELP_KEY variable and binding the key combination to perform
completion.

Signed-off-by: Marc Khouzam <marc.khouzam@montreal.ca>