Disable completion description through env var
By doing
  export COBRA_COMPLETION_DESCRIPTIONS=0
a user can disable completion descriptions.  This can be useful if a
program has not provided an option to disable them.  It also makes it
much simpler for users to disable them when shell completion is
installed through a package manager and the user is not aware of the
mechanics.

Note that using this variable will disable completion descriptions for
ALL programs using Cobra.  This should be sufficient as the expectation
is that users that don't want descriptions won't want them at all.

Signed-off-by: Marc Khouzam <marc.khouzam@montreal.ca>